requirements.txt
acres==0.1.0
asgiref==3.8.1
boto3==1.35.74
botocore==1.35.74
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
Django==5.1.4
django-extensions==3.2.3
django-storages==1.14.4
etelemetry==0.3.1
filelock==3.16.1
gunicorn==23.0.0
httplib2==0.22.0
icalendar==6.1.0
idna==3.10
isodate==0.6.1
jmespath==1.0.1
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
prov==2.0.1
psutil==6.1.0
puremagic==1.28
pycparser==2.22
pydot==3.0.3
PyMuPDF==1.24.14
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
s3transfer==0.10.4
scipy==1.14.1
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.2
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
